Add test that runs the hello configuration twice in a row.
Make certain that the intermediate results and aggregates are cleared
between runs.
Make certain that the intermediate results have no more than the
expected number of results reported therein.
Make certain that the aggregate results have the right number of
results reported therein.
Make certain that the last latency of the first configuration execution
occurs prior to the first latency of the second configuration execution